splitting json.rs in files. Not compiling
